+++ server/publications.js
@@ -39,12 +39,27 @@ Meteor.publish('upvotedPosts', function(userIdOrSlug) {
   var findBySlug = Meteor.users.findOne({slug: userIdOrSlug});
   var user = typeof findById !== 'undefined' ? findById : findBySlug;
   var upvotedPostIds = _.pluck(user.profile.upvotedPosts, 'itemId');
-  if (this.userId = user._id) { // only publish upvoted posts when users are browsing their own profile
-    return Posts.find({_id: {$in: upvotedPostIds}});
+  return Posts.find({_id: {$in: upvotedPostIds}});
+});
+
+Meteor.publish('userComments', function(userIdOrSlug) {
+  var findById = Meteor.users.findOne(userIdOrSlug);
+  var findBySlug = Meteor.users.findOne({slug: userIdOrSlug});
+  var user = typeof findById !== 'undefined' ? findById : findBySlug;
+  return Comments.find({_id: {$in: user.profile.comments}});
+});
+
+Meteor.publish('commentedPosts', function(userIdOrSlug) {
+  var findById = Meteor.users.findOne(userIdOrSlug);
+  var findBySlug = Meteor.users.findOne({slug: userIdOrSlug});
+  var user = typeof findById !== 'undefined' ? findById : findBySlug;
+  var comments = Comments.find({_id: {$in: user.profile.comments}});
+  if(!!comments.count()){ // there might not be any comments
+    var commentedPostIds = _.pluck(comments.fetch(), 'postId');
+    return Posts.find({_id: {$in: commentedPostIds}});
   }
   return [];
 });
-
 
 // Publish authors of the current post and its comments